[refactor] ApiResponse 파일위치 리팩토링
