+++ src/main/java/com/ceos19/springboot/comment/controller/CommentController.java
@@ -6,10 +6,8 @@ import com.ceos19.springboot.comment.dto.CommentResponseDto;
 import com.ceos19.springboot.comment.dto.CommentUserRequestDto;
 import com.ceos19.springboot.comment.service.CommentService;
 import com.ceos19.springboot.commentlike.service.CommentLikeService;
-import com.ceos19.springboot.exception.ApiResponse;
+import com.ceos19.springboot.global.ApiResponse;
 import com.ceos19.springboot.post.domain.Post;
-import com.ceos19.springboot.post.dto.PostCommentResponseDto;
-import com.ceos19.springboot.post.dto.PostResponseDto;
 import com.ceos19.springboot.post.service.PostService;
 import com.ceos19.springboot.users.domain.Users;
 import com.ceos19.springboot.users.service.UserService;
